+++ tools/accuracy_checker/openvino/tools/accuracy_checker/launcher/openvino_launcher.py
@@ -46,7 +46,7 @@ from .launcher import Launcher
 
 
 format_map = {
-      'f32': np.float32, 'i32': np.int32, 'i64': np.int64,
+      'f32': np.float32, 'i32': np.int32, 'i64': int,
       'fp16': np.float16, 'f16': np.float16, 'i16': np.int16, 'u16': np.uint16,
       'i8': np.int8, 'u8': np.uint8, 'boolean': np.uint8
 }
